refactor : BaseEntity 위치 수정
