+++ src/main/java/com/uos/all4runner/domain/entity/common/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.uos.all4runner.common;
+package com.uos.all4runner.domain.entity.common;
 
 import java.time.Instant;
 import java.util.UUID;
