+++ src/main/java/com/uos/all4runner/domain/entity/common/ToiletDrinkVO.java
@@ -1,4 +1,4 @@
-package com.uos.all4runner.common;
+package com.uos.all4runner.domain.entity.common;
 
 import jakarta.persistence.Embeddable;
