+++ src/main/java/com/uos/all4runner/domain/entity/network/Network.java
@@ -1,7 +1,7 @@
 package com.uos.all4runner.domain.entity.network;
 
 
-import com.uos.all4runner.common.ToiletDrinkVO;
+import com.uos.all4runner.domain.entity.common.ToiletDrinkVO;
 import com.uos.all4runner.constant.LinkType;
 import com.uos.all4runner.domain.entity.accountnetwork.AccountNetwork;
